fix: Kill Multithread Bug in Receiving
A bug that the main thread misreads shared data is dead, now that they're mutually excluded even when being read.
etc: Type inference keyword 'auto', literal for unsigned 'u' diminished.
Usage of '__debugbreak()' has been reduced.
Some of 'uint32_t' were replaced with 'uint16_t'.
ASSERT macro gives a hint on a function's precondition.
Identifiers, variable names were renamed to ones with a postpositive adjective, which gives us readability.  For instance, 'mTitleTextLabel' and 'mStartButtonTextLabel' were changed to 'mTextLabelForTitle' and 'mTextLabelForStartButton' respectively.
Access to a value within std::unordered_map by operator[] is rejected, because it assigns a new value unless it has the value with the key.

+++ dlls/Game/scene/inPlay/InPlay.h
@@ -19,10 +19,10 @@ namespace scene::inPlay
 	private:
 		void setScene( const ::scene::inPlay::ID nextInPlaySceneID );
 		static bool IsInstantiated;
-		uint32_t mFPS_;
+		uint16_t mFPS_;
 		sf::RenderWindow& mWindow_;
-		std::unique_ptr< ::scene::inPlay::IScene > mCurrentScene;
-		std::unique_ptr< ::scene::inPlay::IScene > mOverlappedScene;
+		std::unique_ptr<::scene::inPlay::IScene> mCurrentScene;
+		std::unique_ptr<::scene::inPlay::IScene> mOverlappedScene;
 		sf::RectangleShape mBackground;
 	};
 }
